+++ CHANGELOG.md
@@ -30,6 +30,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - `package.version` for QuestPackages and -Templates
 - BlockSelector now supports variables when used in events, conditions, objectives
 - Variable support for all lists in events objectives and conditions for things like events conditions items and co.
+- `Menu Item` are now in the `menu_items` section, allowing references from multiple Menus
 ### Changed
 - Spigot is no longer supported, paper is now required 
 - message.yml file was deleted and instead the lang folder now contains all translations